Fixed missing  in latest migration

+++ db/migrate/20180830053921_change_price_datatype_to_integer2.rb
@@ -1,5 +1,5 @@
 class ChangePriceDatatypeToInteger2 < ActiveRecord::Migration[5.2]
   def change
-    change_column :products, :price, :float
+    change_column :products, :price, :integer
   end
 end
